--- src/main/java/com/blogproject/blogproject/BlogprojectApplication.java
@@ -1,13 +1,0 @@
-package com.blogproject.blogproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BlogprojectApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BlogprojectApplication.class, args);
-    }
-
-}
