--- src/main/java/com/blogproject/category/exception/CategoryNotFoundException.java
@@ -1,8 +1,0 @@
-package com.blogProject.category.exception;
-
-public class CategoryNotFoundException extends RuntimeException {
-
-  public CategoryNotFoundException(String message) {
-    super(message);
-  }
-}
