--- src/main/java/com/blogproject/category/exception/NameAlreadyExistsException.java
@@ -1,8 +1,0 @@
-package com.blogProject.category.exception;
-
-public class NameAlreadyExistsException extends RuntimeException {
-
-  public NameAlreadyExistsException(String message) {
-    super(message);
-  }
-}
