+++ src/main/java/com/blogproject/common/category/dto/CategoryDto.java
@@ -1,0 +1,20 @@
+package com.blogProject.common.category.dto;
+
+import javax.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CategoryDto {
+
+  @NotBlank(message = "카테고리 이름은 필수입니다.")
+  private String categoryName;
+
+}
