+++ src/main/java/com/blogproject/common/category/exception/CategoryNotFoundException.java
@@ -1,0 +1,8 @@
+package com.blogProject.common.category.exception;
+
+public class CategoryNotFoundException extends RuntimeException {
+
+  public CategoryNotFoundException(String message) {
+    super(message);
+  }
+}
