+++ src/main/java/com/blogproject/common/category/exception/NameAlreadyExistsException.java
@@ -1,0 +1,8 @@
+package com.blogProject.common.category.exception;
+
+public class NameAlreadyExistsException extends RuntimeException {
+
+  public NameAlreadyExistsException(String message) {
+    super(message);
+  }
+}
