+++ src/main/java/com/blogproject/common/post/exception/PostNotfoundException.java
@@ -1,0 +1,8 @@
+package com.blogProject.common.post.exception;
+
+public class PostNotfoundException extends RuntimeException {
+
+  public PostNotfoundException(String message) {
+    super(message);
+  }
+}
