--- src/main/java/com/blogproject/post/exception/PostNotfoundException.java
@@ -1,8 +1,0 @@
-package com.blogProject.post.exception;
-
-public class PostNotfoundException extends RuntimeException {
-
-  public PostNotfoundException(String message) {
-    super(message);
-  }
-}
